aws-data-pipeline/requirements.txt
pymysql
pymssql
snowflake-connector-python